refactor: 불필요 라인 제거. code linting파일 추가
