+++ .editorconfig
@@ -1,0 +1,4 @@
+root = true
+
+[*]
+insert_final_newline=true
